Create BadNeighbour Topcoder Problem.cpp
Problem Statement : 
The old song declares "Go ahead and hate your neighbour", and the residents of Bonneville have taken those words to heart. Every resident hates his next-door neighbours on both sides. Nobody is willing to live farther away from the town's well than his neighbours, so the town has been arranged in a big circle around the well. Unfortunately, the town's well is in disrepair and needs to be restored. You have been hired to collect donations for the Save Our Well fund.

Each of the town's residents is willing to donate a certain amount, as specified in the int[] donations, which is listed in clockwise order around the well. However, nobody is willing to contribute to a fund to which his neighbour has also contributed. Next-door neighbours are always listed consecutively in donations, except that the first and last entries in donations are also for next-door neighbours. You must calculate and return the maximum amount of donations that can be collected.